chore: incorrect classifier usage
The Development Status classifier is _not_ a proxy for the version status. It's an indicator for how stable the _project as a whole_ is. That's why the don't match dev/alpha/beta/rc/final/post, but instead have things like 6 - Mature and 7 - Inactive. Coverage is a 5 - Production / Stable. It is, as a project, Production / Stable even in development and beta releases. The dev release tells you it's a dev release (of a stable project).

Signed-off-by: Henry Schreiner <henryfs@princeton.edu>